+++ golang/lambda/example_lambda_args.go
@@ -1,0 +1,11 @@
+/*
+ * Copyright (c) 2021 - present Kurtosis Technologies LLC.
+ * All Rights Reserved.
+ */
+
+package lambda
+
+type ExampleLambdaArgs struct {
+	// Indicates the log level for this Lambda implementation
+	LogLevel string `json:"logLevel"`
+}