+++ docs/changelog.md
@@ -1,4 +1,12 @@
 # TBD
+### Features
+* Added `ExampleLambda` which is an example of how to implement a Kurtosis Lambda
+* Added `ExampleLambdaConfigurator` and `ExampleLambdaArgs` which are used to start the `ExampleLambda`implementation 
+* Added the main go file for startup the Lambda
+* Added Dockerfile with build and run configurations
+* Added Circle CI configuration file
+* Added golang directory and go mod file
+* Added shell script to execute kurtosis release repository script for this repository
 
 #0.1.0
 * Init commit